+++ SalesforceToolKit/Localizables/es.lproj/Localizable.strings
@@ -20,3 +20,5 @@
 "Others" = "Otros";
 "About" = "Acerca de";
 "Help" = "Ayuda";
+"default" = "Por defecto";
+"chrome" = "Google Chrome";
